requirements.txt
opencv-python
mediapipe
ffmpeg-python
